bundles/bme280/files/requirements.txt
paho-mqtt
RPi.bme280
